+++ napari/_qt/qt_chunk_receiver.py
@@ -4,7 +4,7 @@ import logging
 
 from qtpy.QtCore import QObject, Signal
 
-from ..components.chunk import chunk_loader
+from ..components.experimental.chunk import chunk_loader
 from ..utils.events import EmitterGroup, Event, EventEmitter
 
 LOGGER = logging.getLogger('napari.async')
