+++ napari/components/experimental/chunk/_request.py
@@ -5,8 +5,8 @@ from typing import Optional, Tuple
 
 import numpy as np
 
-from ...layers.base.base import Layer
-from ...types import ArrayLike, Dict
+from ....layers.base.base import Layer
+from ....types import ArrayLike, Dict
 from ._utils import get_data_id
 
 LOGGER = logging.getLogger("napari.async")
