+++ napari/components/experimental/chunk/_utils.py
